1.0/eureka/requirements.txt
Django==2.2.4
django-cors-headers==3.2.1
djangorestframework==3.11.0
requests==2.22.0
PyMySQL==0.9.3
